+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [Also based on](https://github.com/conventional-changelog/standard-version/blob/master/CHANGELOG.md) so decending.
 
 ## [0.4.0] - 2021-12-29
+### Removed
+- removes the wordplay class since it was unused
+
 ### Changed
 - refactors the verbs into their own list variables